apps/mcp-server/pyproject.toml
[project]
name = "vivian-mcp-server"
version = "0.1.0"
description = "MCP server for Vivian household agent tools"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
vivian-mcp = "vivian_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vivian_mcp"]
